attempts to fix TR-214
