+++ public/canjs/controls/timeline-configuration/timeline-configuration.js
@@ -174,8 +174,7 @@ export class TimelineConfiguration extends StacheElement {
         <div class="{{^ eq(this.routeData.showSettings, "TIMING")}}hidden{{/}}">
             ${GOBACK_BUTTON}
             <timing-calculation 
-                jiraHelpers:from="this.jiraHelpers"
-                issueTimingCalculations:to="this.issueTimingCalculations"></timing-calculation>
+                jiraHelpers:from="this.jiraHelpers"></timing-calculation>
         </div>
         
 
@@ -198,9 +197,6 @@ export class TimelineConfiguration extends StacheElement {
         return routeData;
       }
     },
-
-    // from children
-    issueTimingCalculations: null,
 
     // VALUES DERIVING FROM THE `jql`
