+++ public/canjs/controls/timeline-configuration/timing-calculation/timing-calculation.js
@@ -8,6 +8,8 @@ import routeData from "../../../routing/route-data.js";
 const selectStyle = "bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 p-2 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
 
 
+import {getTimingLevels} from "./helpers.js";
+
 const DEFAULT_CALCULATION_METHOD = "widestRange";
 
 export class TimingCalculation extends StacheElement {
@@ -42,10 +44,7 @@ export class TimingCalculation extends StacheElement {
             }
         },
         get jiraIssueHierarchyPromise(){
-            return getSimplifiedIssueHierarchy({
-                isLoggedIn: this.jiraHelpers.hasValidAccessToken(),
-                jiraHelpers: this.jiraHelpers,
-            }) 
+            return this.routeData.simplifiedIssueHierarchyPromise;
         },
         issueHierarchy: {
             async(resolve){
@@ -58,20 +57,6 @@ export class TimingCalculation extends StacheElement {
             } else {
                 const allLevels = getTimingLevels(this.issueHierarchy, this.routeData.timingCalculations);
                 return allLevels.slice(0, allLevels.length - 1);
-            }
-        },
-        get issueTimingCalculations(){
-            if(!this.issueHierarchy) {
-                return [];
-            } else {
-                const allLevels = getTimingLevels(this.issueHierarchy, this.routeData.timingCalculations);
-                return allLevels.map( level => {
-                    return {
-                        type: level.type,
-                        hierarchyLevel: level.hierarchyLevel,
-                        calculation: level.calculations.find( (level) => level.selected).calculation
-                    }
-                })
             }
         }
     }
@@ -124,89 +109,8 @@ function getIssueHierarchy(types){
 */
 
 
-export const calculationKeysToNames = {
-    parentFirstThenChildren: function(parent, child){
-        return `From ${parent.type}, then ${child.plural}`
-    },
-    childrenOnly: function(parent, child){
-        return `From ${child.plural}`
-    },
-    childrenFirstThenParent: function(parent, child){
-        return `From ${child.plural}, then ${parent.type}`
-    },
-    widestRange: function(parent, child){
-        return `From ${parent.type} or ${child.plural} (earliest to latest)`
-    },
-    parentOnly: function(parent, child){
-        return `From ${parent.type}`
-    }
-}
-
-function createBaseLevels(issueHierarchy) {
-    return issueHierarchy.map((issue)=> {
-        return {
-            type: issue.name,
-            source: issue,
-            plural: issue.name+"s",
-            hierarchyLevel: issue.hierarchyLevel
-        }
-    });
-}
-
-function calculationsForLevel(parent, child, selected, last){
-    if(!last) {
-        return Object.keys(calculationKeysToNames).map( calculationName => {
-            return {
-                parent: parent.type,
-                child: child.type,
-                calculation: calculationName,
-                name:  calculationKeysToNames[calculationName](parent, child),
-                selected: selected ? selected === calculationName : "widestRange" === calculationName
-            }
-        })
-    } else {
-        return [{
-            parent: parent.type,
-            child: null,
-            calculation: "parentOnly",
-            name:  calculationKeysToNames.parentOnly(parent),
-            selected: true
-        }];
-    }
-}
-
-/*
-return {
-    child: issueTypeName, 
-    parent: issueType.type, 
-    calculation: calculationName, name: calculationKeysToNames[calculationName](issueType, typeToIssueType[issueTypeName]) }
-*/
-
-/**
-* 
-* @param {TimingCalculationsMap} issueTypeMap 
-* @param {string} primaryIssueType 
-* @param {Array<TimingCalculation>} timingCalculations 
-* @returns 
-*/
-function getTimingLevels(issueHierarchy, timingCalculations){
 
 
-    const baseLevels = createBaseLevels(issueHierarchy);
-
-    return baseLevels.map( (level, i)=> {
-        const child = baseLevels[i+1];
-        const isLast = i === baseLevels.length - 1;
-
-        return {
-            ...level,
-            childType: child ? child.type : null,
-            calculations: calculationsForLevel(level, child, timingCalculations[level.type], isLast)
-        }
-    });
-    
-   
-}
 
 
 customElements.define("timing-calculation", TimingCalculation);
